requirements.txt
email_validator==1.3.1
Flask==2.2.3
Flask_Migrate==4.0.4
flask_sqlalchemy==3.0.3
iyzipay==1.0.32
Pillow==9.5.0
pytest==7.3.1
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.25.1
responses==0.23.1
